@@ -4,11 +4,12 @@
 
 'use strict';
 
-var errors = require('./components/errors');
 var express = require('express');
 var config = require('./config/environment');
+import errors from './components/errors';
+import path from 'path';
 
-module.exports = function(app) {
+export default function(app) {
 
   // Insert routes below
   app.use('/api/posts', require('./api/post'));
@@ -17,6 +18,7 @@ module.exports = function(app) {
   app.use('/api/users', require('./api/user'));
   app.use('/api/classyear', require('./api/classyear'));
   app.use('/api/smallgroup', require('./api/smallgroup'));
+  app.use('/api/attendance', require('./api/attendance'));
 
   app.use('/uploads', express.static(config.imageUploadPath));
 
@@ -26,9 +28,9 @@ module.exports = function(app) {
   app.route('/:url(api|auth|components|app|bower_components|assets|uploads)/*')
    .get(errors[404]);
 
-  //All other routes should redirect to the index.html
+  // All other routes should redirect to the index.html
   app.route('/*')
-    .get(function(req, res) {
-      res.sendfile(app.get('appPath') + '/index.html');
+    .get((req, res) => {
+      res.sendFile(path.resolve(app.get('appPath') + '/index.html'));
     });
-};
+}
