@@ -2,22 +2,23 @@
 
 angular.module('observatory3App')
   .controller('MentorCtrl', function ($scope, $stateParams, $http, Auth, User, $location, notify) {
-    function updateUnverified(){
-        $http.get("/api/users/unverified")
-        	.success(function(unverifiedUsers){
-        		$scope.unverifiedUsers = unverifiedUsers;
+    var updateUnverified = function(){
+        $http.get('/api/attendance/unverified/today')
+        	.success(function(unverifiedSubmissions){
+        		$scope.unverifiedSubmissions = unverifiedSubmissions;
         	})
         	.error(function(err){
-            notify({ message: "Error Occurred: " + err, classes: ["alert-danger"] });
+                notify({ message: 'Error Occurred: ' + err, classes: ['alert-danger'] });
         	});
-    }
-    $scope.verifyAttendance = function(user){
-    	$http.put("/api/users/" + user._id + "/verifyAttendance")
+    };
+
+    $scope.verifyAttendance = function(submission){
+    	$http.put('/api/attendance/' + submission._id + '/verify')
             .success(function(){
                 updateUnverified();
             })
             .error(function(err){
-                notify({ message: "Error Occurred: " + err, classes: ["alert-danger"] });
+                notify({ message: 'Error Occurred: ' + err, classes: ['alert-danger'] });
             });
     };
     updateUnverified();
