@@ -16,7 +16,7 @@ angular.module('observatory3App')
     $scope.isMentor = Auth.isMentor;
     $scope.getCurrentUser = Auth.getCurrentUser;
 
-		$scope.logout = function() {
+	$scope.logout = function() {
       Auth.logout();
       $location.path('/login');
     };
