@@ -10,8 +10,8 @@ var ProjectSchema = new Schema({
   repositoryType: {type: String, default: 'github'},
   repositories: [String],
   websiteUrl: String,
-  githubUsername: {type: String, index: true},
-  githubProjectName: {type: String, index: true},
+  githubUsername: {type: String, lowercase:true, index: true},
+  githubProjectName: {type: String, lowercase:true, index: true},
   lastChecked: {type: Date},
   authors: [String],
   photos: [String],
